long -> int refactor. Impl load/store.

+++ src/main/java/com/neilcochran/Main.java
@@ -14,7 +14,7 @@ public class Main {
      * @param args Main args are currently not used
      */
     public static void main(String[] args) {
-        long[] programData = {
+        int[] programData = {
             0b00001011100000000000000000000100,
             0b00000000000000000000000000000001,
             0b00000000000000000000000000000010,
